#277 Rename the files so they represent the services
