+++ services/energy-monitor/src/index.ts
@@ -1,7 +1,7 @@
 import { LoggerService, MqttService } from "@powerpi/common";
 import Container from "./container";
 import EnergyMonitorArgumentsService from "./services/EnergyMonitorArgumentService";
-import EnergyMonitorService from "./services/monitor";
+import EnergyMonitorService from "./services/EnergyMonitorService";
 
 function start() {
     const args = Container.get(EnergyMonitorArgumentsService);
