+++ services/energy-monitor/src/models/N3rgyData.ts
