+++ services/energy-monitor/src/services/ConfigService.ts
